+++ point.go
@@ -3,6 +3,6 @@ package convtree
 type Point struct {
 	X       float64
 	Y       float64
-	Weight  int
+	Weight  float64
 	Content interface{}
 }
